requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
joblib
plotly
langgraph
groq
python-dotenv